LayoutTests/imported/w3c/web-platform-tests/trusted-types/Node-multiple-arguments-expected.txt

CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating 'window.trustedTypes.createPolicy')

Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'window.trustedTypes.createPolicy')
